wip; move scheduler client to server package
