+++ airbyte-container-orchestrator/build.gradle
@@ -16,7 +16,7 @@ dependencies {
     implementation project(':airbyte-db:db-lib')
     implementation project(':airbyte-json-validation')
     implementation project(':airbyte-protocol:protocol-models')
-    implementation project(':airbyte-scheduler:scheduler-persistence')
+    implementation project(':airbyte-persistence:job-persistence')
     implementation project(':airbyte-workers')
     implementation project(':airbyte-metrics:metrics-lib')
