--- airbyte-scheduler/client/readme.md
@@ -1,6 +1,0 @@
-# airbyte-scheduler:client
-
-Java clients for submitting Jobs.
-
-## Key Files
-* `SynchronousSchedulerClient` - interface for scheduling _synchronous_ jobs.
