--- airbyte-scheduler/scheduler-models/build.gradle
@@ -1,10 +1,0 @@
-plugins {
-    id "java-library"
-}
-
-dependencies {
-    implementation project(':airbyte-config:config-models')
-    implementation project(':airbyte-protocol:protocol-models')
-}
-
-Task publishArtifactsTask = getPublishArtifactsTask("$rootProject.ext.version", project)
