--- airbyte-scheduler/scheduler-models/readme.md
@@ -1,8 +1,0 @@
-# airbyte-scheduler:scheduler-models
-
-This module declares models that belong to the Scheduler.
-
-These models are generated in the same way as `airbyte-config:models`. See that module for reference.
-
-## Key Files
-* `Job.java` and `Attempt.java` are the top-level models in this schema.
