--- airbyte-scheduler/scheduler-persistence/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-    id "java-library"
-}
-
-dependencies {
-    implementation 'io.sentry:sentry:6.1.0'
-
-    implementation project(':airbyte-analytics')
-    implementation project(':airbyte-commons-docker')
-    implementation project(':airbyte-config:config-models')
-    implementation project(':airbyte-config:config-persistence')
-    implementation project(':airbyte-db:db-lib')
-    implementation project(':airbyte-db:jooq')
-    implementation project(':airbyte-json-validation')
-    implementation project(':airbyte-notification')
-    implementation project(':airbyte-oauth')
-    implementation project(':airbyte-protocol:protocol-models')
-    implementation project(':airbyte-scheduler:scheduler-models')
-
-    testImplementation libs.flyway.core
-    testImplementation libs.platform.testcontainers.postgresql
-    testImplementation project(':airbyte-test-utils')
-}
-
-Task publishArtifactsTask = getPublishArtifactsTask("$rootProject.ext.version", project)
