+++ airbyte-server/build.gradle
@@ -15,9 +15,6 @@ dependencies {
     implementation project(':airbyte-notification')
     implementation project(':airbyte-oauth')
     implementation project(':airbyte-protocol:protocol-models')
-    implementation(project(':airbyte-scheduler:client')) {
-        exclude module: 'airbyte-workers'
-    }
     implementation project(':airbyte-persistence:job-persistence')
     implementation(project(':airbyte-workers')) {
         // Temporary hack to avoid dependency conflicts
